refactor: separate typings and declarations
